backend/requirements.txt
aiohttp==3.9.0b0
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.8.0
attrs==23.2.0
Cython==3.0.10
Django==5.0.3
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-ninja==1.1.0
django-pandas==0.6.7
frozenlist==1.4.1
idna==3.7
joblib==1.4.2
multidict==6.0.5
numpy==1.26.4
pandas==2.2.1
pillow==10.3.0
psycopg==3.1.18
psycopg-binary==3.1.18
pydantic==2.6.4
pydantic_core==2.16.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.4.2
scikit-surprise==1.1.3
scipy==1.13.0
setuptools==70.0.0
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.5.0
typing_extensions==4.10.0
tzdata==2024.1
wheel==0.43.0
yarl==1.9.4